+++ src/app/shared/components/cvss-entry/cvss-entry.component.ts
@@ -1,9 +1,10 @@
 import { Component, Input, OnInit, Optional } from '@angular/core';
 import { ICVSSEntry, ThreatSeverities } from '../../../model/threat-model';
+import { Clipboard } from '@angular/cdk/clipboard';
 
 import Cvss from 'cvss-calculator';
 import { DataService } from '../../../util/data.service';
-import { MyCVSSEntry } from '../../../util/dialog.service';
+import { DialogService, MyCVSSEntry } from '../../../util/dialog.service';
 
 @Component({
   selector: 'app-cvss-entry',
@@ -18,12 +19,24 @@ export class CvssEntryComponent implements OnInit {
   }
   public set Score(val: number) { this.entry.Score = val; }
 
-  constructor(@Optional() public data: MyCVSSEntry, public dataService: DataService) {
+  public get Vector(): string {
+    return CvssEntryComponent.GetVector(this.entry);
+  }
+  public set Vector(val: string) {
+    this.SetVector(val);
+  }
+
+  constructor(@Optional() public data: MyCVSSEntry, public dataService: DataService, private dialog: DialogService, private clipboard: Clipboard) {
     if (data) this.entry = data.Value;
   }
 
   ngOnInit(): void {
     if (!this.Score) this.CalcScore();
+    if (!this.entry?.Notes) this.entry.Notes = {};
+  }
+
+  public CopyVector() {
+    this.clipboard.copy(this.Vector);
   }
 
   public OpenCVSS() {
@@ -37,6 +50,97 @@ export class CvssEntryComponent implements OnInit {
       const cvss = new Cvss(vec);
       this.Score = cvss.getBaseScore();
     }
+  }
+
+  public SetVector(val: string) {
+    if (val) {
+      let valid = true;
+      const validMetrics = ['AV', 'AC', 'PR', 'UI', 'S', 'C', 'I', 'A', 'CVSS'];
+      const removeMetric = (metric: string) => {
+        const index = validMetrics.indexOf(metric);
+        if (index >= 0) validMetrics.splice(index, 1);
+      };
+
+      const res = {};
+      const metrics = val.split('/');
+      metrics.forEach(metric => {
+        if (valid) {
+          const parts = metric.split(':');
+          if (parts.length == 2) {
+            if (validMetrics.includes(parts[0])) {
+              switch (parts[0]) {
+                case 'AV':
+                  if (['N', 'AN', 'L', 'P'].includes(parts[1])) { res[parts[0]] = parts[1]; removeMetric(parts[0]); }
+                  else valid = false;
+                  break;
+                case 'AC':
+                  if (['L', 'H'].includes(parts[1])) { res[parts[0]] = parts[1]; removeMetric(parts[0]); }
+                  else valid = false;
+                  break;
+                case 'PR':
+                  if (['N', 'L', 'H'].includes(parts[1])) { res[parts[0]] = parts[1]; removeMetric(parts[0]); }
+                  else valid = false;
+                  break;
+                case 'UI':
+                  if (['N', 'R'].includes(parts[1])) { res[parts[0]] = parts[1]; removeMetric(parts[0]); }
+                  else valid = false;
+                  break;
+                case 'S':
+                  if (['U', 'C'].includes(parts[1])) { res[parts[0]] = parts[1]; removeMetric(parts[0]); }
+                  else valid = false;
+                  break;
+                case 'C':
+                  if (['N', 'L', 'H'].includes(parts[1])) { res[parts[0]] = parts[1]; removeMetric(parts[0]); }
+                  else valid = false;
+                  break;
+                case 'I':
+                  if (['N', 'L', 'H'].includes(parts[1])) { res[parts[0]] = parts[1]; removeMetric(parts[0]); }
+                  else valid = false;
+                  break;
+                case 'A':
+                  if (['N', 'L', 'H'].includes(parts[1])) { res[parts[0]] = parts[1]; removeMetric(parts[0]); }
+                  else valid = false;
+                  break;
+                case 'CVSS':
+                  if (['2.0', '3.0', '3.1'].includes(parts[1])) { res['Version'] = parts[1]; removeMetric(parts[0]); }
+                  else valid = false;
+                  break;
+              
+                default:
+                  valid = false;
+                  break;
+              }
+            }
+            else {
+              valid = false;
+            }
+          }
+          else {
+            valid = false;
+          }
+        }
+      });
+
+      if (valid) {
+        Object.keys(res).forEach(key => {
+          this.entry[key] = res[key];
+        });
+        this.CalcScore();
+      }
+    }
+  }
+
+  public OpenNotes(metric: string) {
+    if (this.entry.Notes[metric] == null) this.entry.Notes[metric] = [];
+    const notes = this.entry.Notes[metric];
+    this.dialog.OpenNotesDialog(notes, true, false, true, true);
+  }
+
+  public GetNotesCountOfMetric(metric: string): number {
+    if (this.entry.Notes[metric]) {
+      return this.entry.Notes[metric].length;
+    }
+    return 0;
   }
 
   public static ToThreatSeverity(score: number) {
@@ -59,7 +163,7 @@ export class CvssEntryComponent implements OnInit {
       let vec = 'CVSS:' + (entry.Version ? entry.Version : '3.1');
       if (entry.Vector) return entry.Vector.includes('CVSS') ? entry.Vector : (vec + '/' + entry.Vector);
       Object.keys(entry).forEach(k => {
-        if (entry[k] && typeof entry[k] === 'string') vec += '/' + k + ':' + entry[k];
+        if (entry[k] && typeof entry[k] === 'string' && k.length <= 4) vec += '/' + k + ':' + entry[k];
       });
       return vec;
     }